backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.2
django-cors-headers==4.7.0
nltk==3.9.1
textblob==0.19.0
pandas==2.2.3
scikit-learn==1.6.1
requests==2.31.0
google-generativeai==0.8.4
